refactor(trusty-common): split semantic_consolidation.rs into submodules under 500-SLOC cap (#607)
Replaces the 615-SLOC semantic_consolidation.rs monolith with a
semantic_consolidation/ directory module:

  semantic_consolidation/
    mod.rs          — module doc, re-exports, all tests (228 SLOC)
    types.rs        — SemanticConsolidationConfig, ConsolidationAction,
                      ConsolidationResult, CanonicalDrawer, inference_available,
                      parse_consolidation_actions, batch_cache_key,
                      apply_actions_to_result, wire types (160 SLOC)
    inference.rs    — Inference trait, OpenRouterInference, OllamaInference,
                      MockInference, CONSOLIDATION_PROMPT_SYSTEM (185 SLOC)
    consolidator.rs — SemanticConsolidator struct + consolidate() (65 SLOC)

Zero API change: all pub items remain re-exported from memory_core::semantic_consolidation.
